libpayload: Fix inl() return type mismatch
Change `inl()` return type from `unsigned long` to `unsigned int` to
match the function definition and ensure consistency across platforms.

BUG=b:242829490
TEST=Compiled successfully in 32-bit and 64-bit modes.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/81961
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ payloads/libpayload/include/x86/arch/io.h
@@ -90,7 +90,7 @@ static inline __attribute__((always_inline)) void write64(volatile void *addr, u
 
 static inline unsigned int inl(int port)
 {
-	unsigned long val;
+	unsigned int val;
 	__asm__ __volatile__("inl %w1, %0" : "=a"(val) : "Nd"(port));
 	return val;
 }
